Move selfdev worker contract into core
